Drop residential address validation
There are too many edge cases for the residential address validation, not to mention inconsistencies on the part of LocationIQ in reporting place type data.
For now, it's easier to just check that a provided address contains a building number.
Of course, this introduces the issue of allowing users to be able to enter the address of some random McDonald's or something and try to pass it off as a place to stay. Thus, I will open residential address validation as an issue that perhaps someone with more interest and skill in backend logic can tackle.